Add Logical/ANSI Segments

+++ Giselle.Net.EtherNetIP/DataProcessor.cs
@@ -186,7 +186,7 @@ namespace Giselle.Net.EtherNetIP
 
         }
 
-        public virtual byte[] ReadBytes(int length)
+        public virtual byte[] ReadBytes(long length)
         {
             var bytes = new byte[length];
             this.ReadBytes(bytes);
